requirements.txt
Django==5.0.4
djangorestframework~=3.15.1
djangorestframework-simplejwt
drf-spectacular
psycopg2-binary
pip~=24.0
attrs~=23.2.0
api~=0.0.7
PyYAML~=6.0.1
asgiref~=3.8.1
sqlparse~=0.4.4
requests~=2.31.0
referencing~=0.34.0
jsonschema~=4.21.1
idna~=3.6
setuptools~=65.5.0
uritemplate~=4.1.1
inflection~=0.5.1
PyJWT~=2.8.0